+++ mayan/apps/document_indexing/migrations/0024_auto_20210822_0052.py
@@ -1,0 +1,15 @@
+from django.db import migrations
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('document_indexing', '0023_auto_20210821_2059'),
+    ]
+
+    operations = [
+        migrations.AlterUniqueTogether(
+            name='indexinstancenode',
+            unique_together={('parent', 'value')},
+        ),
+    ]